Struct google_genomics1_beta2::Annotation

pub struct Annotation {
    pub info: Option<HashMap<String, Vec<String>>>,
    pub name: Option<String>,
    pub transcript: Option<Transcript>,
    pub variant: Option<VariantAnnotation>,
    pub annotation_set_id: Option<String>,
    pub position: Option<RangePosition>,
    pub type_: Option<String>,
    pub id: Option<String>,
}

An annotation describes a region of reference genome. The value of an annotation may be one of several canonical types, supplemented by arbitrary info tags. A variant annotation is represented by one or more of these canonical types. An annotation is not inherently associated with a specific sample or individual (though a client could choose to use annotations in this way). Example canonical annotation types are 'Gene' and 'Variant'.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

search annotations (none)

patch annotations (request|response)

create annotations (request|response)

delete annotations (none)

update annotations (request|response)

get annotations (response)

batch create annotations (none)

Fields

info

A map of additional data for this annotation.

name

The display name of this annotation.

transcript

A transcript value represents the assertion that a particular region of the reference genome may be transcribed as RNA. An alternative splicing pattern would be represented as a separate transcript object. This field is only set for annotations of type TRANSCRIPT.

variant

A variant annotation, which describes the effect of a variant on the genome, the coding sequence, and/or higher level consequences at the organism level e.g. pathogenicity. This field is only set for annotations of type VARIANT.

annotation_set_id

The ID of the containing annotation set.

position

The position of this annotation on the reference sequence.

type_

The data type for this annotation. Must match the containing annotation set's type.

id

The generated unique ID for this annotation.