Struct google_genomics1_beta2::AnnotationMethods

pub struct AnnotationMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on annotation resources. It is not used directly, but through the Genomics hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_genomics1_beta2 as genomics1_beta2;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use genomics1_beta2::Genomics;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Genomics::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `batch_create(...)`, `create(...)`, `delete(...)`, `get(...)`, `patch(...)`, `search(...)` and `update(...)`
// to build up your call.
let rb = hub.annotations();

Methods

impl<'a, C, A> AnnotationMethods<'a, C, A>

fn update(&self, request: Annotation, annotation_id: &str) -> AnnotationUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Updates an annotation. The update must respect all mutability restrictions and other invariants described on the annotation resource. Caller must have WRITE permission for the associated dataset.

Arguments

request - No description provided.

annotationId - The ID of the annotation set to be updated.

fn get(&self, annotation_id: &str) -> AnnotationGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets an annotation. Caller must have READ permission for the associated annotation set.

Arguments

annotationId - The ID of the annotation set to be retrieved.

fn search(&self, request: SearchAnnotationsRequest) -> AnnotationSearchCall<'a, C, A>

Create a builder to help you perform the following task:

Searches for annotations that match the given criteria. Results are returned ordered by start position. Annotations that have matching start positions are ordered deterministically. Caller must have READ permission for the queried annotation sets.

Arguments

request - No description provided.

fn delete(&self, annotation_id: &str) -> AnnotationDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes an annotation. Caller must have WRITE permission for the associated annotation set.

Arguments

annotationId - The ID of the annotation set to be deleted.

fn patch(&self, request: Annotation, annotation_id: &str) -> AnnotationPatchCall<'a, C, A>

Create a builder to help you perform the following task:

Updates an annotation. The update must respect all mutability restrictions and other invariants described on the annotation resource. Caller must have WRITE permission for the associated dataset. This method supports patch semantics.

Arguments

request - No description provided.

annotationId - The ID of the annotation set to be updated.

fn create(&self, request: Annotation) -> AnnotationCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a new annotation. Caller must have WRITE permission for the associated annotation set.

Arguments

request - No description provided.

fn batch_create(&self, request: BatchCreateAnnotationsRequest) -> AnnotationBatchCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates one or more new annotations atomically. All annotations must belong to the same annotation set. Caller must have WRITE permission for this annotation set. For optimal performance, batch positionally adjacent annotations together.

If the request has a systemic issue, such as an attempt to write to an inaccessible annotation set, the entire RPC will fail accordingly. For lesser data issues, when possible an error will be isolated to the corresponding batch entry in the response; the remaining well formed annotations will be created normally.

Arguments

request - No description provided.