Struct google-genomics1_beta2::Genomics

pub struct Genomics<C, NC, A> {
    // some fields omitted
}

Central instance to access all Genomics related resource activities

Examples

Instantiate a new hub

extern crate hyper;
extern crate "yup-oauth2" as oauth2;
extern crate serde;
extern crate "google-genomics1_beta2" as genomics1_beta2;
use genomics1_beta2::Annotation;
use genomics1_beta2::Result;
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use genomics1_beta2::Genomics;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Genomics::new(hyper::Client::new(), auth);
// As the method needs a request, you would usually fill it with the desired information
// into the respective structure. Some of the parts shown here might not be applicable !
// Values shown here are possibly random and not representative !
let mut req: Annotation = Default::default();
 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.annotations().patch(&req, "annotationId")
             .doit();
 
match result {
    Result::HttpError(err) => println!("HTTPERROR: {:?}", err),
    Result::MissingAPIKey => println!("Missing API Key"),
    Result::MissingToken => println!("Missing Token"),
    Result::Failure(_) => println!("General Failure (Response doesn't print)"),
    Result::FieldClash(clashed_field) => println!("FIELD CLASH: {:?}", clashed_field),
    Result::JsonDecodeError(err) => println!("Json failed to decode: {:?}", err),
    Result::Success(_) => println!("Success (value doesn't print)"),
}

Methods

impl<'a, C, NC, A> Genomics<C, NC, A> where NC: NetworkConnector, C: BorrowMut<Client<NC>>, A: GetToken

fn new(client: C, authenticator: A) -> Genomics<C, NC, A>

fn annotation_sets(&'a self) -> AnnotationSetMethodsBuilder<'a, C, NC, A>

fn annotations(&'a self) -> AnnotationMethodsBuilder<'a, C, NC, A>

fn callsets(&'a self) -> CallsetMethodsBuilder<'a, C, NC, A>

fn datasets(&'a self) -> DatasetMethodsBuilder<'a, C, NC, A>

fn experimental(&'a self) -> ExperimentalMethodsBuilder<'a, C, NC, A>

fn jobs(&'a self) -> JobMethodsBuilder<'a, C, NC, A>

fn readgroupsets(&'a self) -> ReadgroupsetMethodsBuilder<'a, C, NC, A>

fn reads(&'a self) -> ReadMethodsBuilder<'a, C, NC, A>

fn references(&'a self) -> ReferenceMethodsBuilder<'a, C, NC, A>

fn referencesets(&'a self) -> ReferencesetMethodsBuilder<'a, C, NC, A>

fn variants(&'a self) -> VariantMethodsBuilder<'a, C, NC, A>

fn variantsets(&'a self) -> VariantsetMethodsBuilder<'a, C, NC, A>

fn user_agent(&mut self, agent_name: String) -> String

Set the user-agent header field to use in all requests to the server. It defaults to google-api-rust-client/0.0.1.

Returns the previously set user-agent.