Struct google_genomics1_beta2::ImportVariantsResponse

pub struct ImportVariantsResponse {
    pub job_id: Option<String>,
}

The variant data import response.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

import variants variantsets (response)

Fields

job_id

A job ID that can be used to get status information.