Struct google_genomics1_beta2::Read

pub struct Read {
    pub read_group_set_id: Option<String>,
    pub failed_vendor_quality_checks: Option<bool>,
    pub fragment_name: Option<String>,
    pub next_mate_position: Option<Position>,
    pub supplementary_alignment: Option<bool>,
    pub fragment_length: Option<i32>,
    pub id: Option<String>,
    pub alignment: Option<LinearAlignment>,
    pub info: Option<HashMap<String, Vec<String>>>,
    pub duplicate_fragment: Option<bool>,
    pub aligned_quality: Option<Vec<i32>>,
    pub read_number: Option<i32>,
    pub proper_placement: Option<bool>,
    pub read_group_id: Option<String>,
    pub number_reads: Option<i32>,
    pub aligned_sequence: Option<String>,
    pub secondary_alignment: Option<bool>,
}

A read alignment describes a linear alignment of a string of DNA to a reference sequence, in addition to metadata about the fragment (the molecule of DNA sequenced) and the read (the bases which were read by the sequencer). A read is equivalent to a line in a SAM file. A read belongs to exactly one read group and exactly one read group set. Generating a reference-aligned sequence string When interacting with mapped reads, it's often useful to produce a string representing the local alignment of the read to reference. The following pseudocode demonstrates one way of doing this: out = "" offset = 0 for c in read.alignment.cigar { switch c.operation { case "ALIGNMENT_MATCH", "SEQUENCE_MATCH", "SEQUENCE_MISMATCH": out += read.alignedSequence[offset:offset+c.operationLength] offset += c.operationLength break case "CLIP_SOFT", "INSERT": offset += c.operationLength break case "PAD": out += repeat("*", c.operationLength) break case "DELETE": out += repeat("-", c.operationLength) break case "SKIP": out += repeat(" ", c.operationLength) break case "CLIP_HARD": break } } return out Converting to SAM's CIGAR string The following pseudocode generates a SAM CIGAR string from the cigar field. Note that this is a lossy conversion (cigar.referenceSequence is lost). cigarMap = { "ALIGNMENT_MATCH": "M", "INSERT": "I", "DELETE": "D", "SKIP": "N", "CLIP_SOFT": "S", "CLIP_HARD": "H", "PAD": "P", "SEQUENCE_MATCH": "=", "SEQUENCE_MISMATCH": "X", } cigarStr = "" for c in read.alignment.cigar { cigarStr += c.operationLength + cigarMap[c.operation] } return cigarStr

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

search reads (none)

Fields

read_group_set_id

The ID of the read group set this read belongs to. (Every read must belong to exactly one read group set.)

failed_vendor_quality_checks

SAM flag 0x200

fragment_name

The fragment name. Equivalent to QNAME (query template name) in SAM.

next_mate_position

The position of the primary alignment of the (readNumber+1)%numberReads read in the fragment. It replaces mate position and mate strand in SAM. This field will be unset if that read is unmapped or if the fragment only has a single read.

supplementary_alignment

Whether this alignment is supplementary. Equivalent to SAM flag 0x800. Supplementary alignments are used in the representation of a chimeric alignment. In a chimeric alignment, a read is split into multiple linear alignments that map to different reference contigs. The first linear alignment in the read will be designated as the representative alignment; the remaining linear alignments will be designated as supplementary alignments. These alignments may have different mapping quality scores. In each linear alignment in a chimeric alignment, the read will be hard clipped. The alignedSequence and alignedQuality fields in the alignment record will only represent the bases for its respective linear alignment.

fragment_length

The observed length of the fragment, equivalent to TLEN in SAM.

id

The unique ID for this read. This is a generated unique ID, not to be confused with fragmentName.

alignment

The linear alignment for this alignment record. This field will be unset if the read is unmapped.

info

A map of additional read alignment information.

duplicate_fragment

The fragment is a PCR or optical duplicate (SAM flag 0x400)

aligned_quality

The quality of the read sequence contained in this alignment record. alignedSequence and alignedQuality may be shorter than the full read sequence and quality. This will occur if the alignment is part of a chimeric alignment, or if the read was trimmed. When this occurs, the CIGAR for this read will begin/end with a hard clip operator that will indicate the length of the excised sequence.

read_number

The read number in sequencing. 0-based and less than numberReads. This field replaces SAM flag 0x40 and 0x80.

proper_placement

The orientation and the distance between reads from the fragment are consistent with the sequencing protocol (SAM flag 0x2)

read_group_id

The ID of the read group this read belongs to. (Every read must belong to exactly one read group.)

number_reads

The number of reads in the fragment (extension to SAM flag 0x1).

aligned_sequence

The bases of the read sequence contained in this alignment record, without CIGAR operations applied. alignedSequence and alignedQuality may be shorter than the full read sequence and quality. This will occur if the alignment is part of a chimeric alignment, or if the read was trimmed. When this occurs, the CIGAR for this read will begin/end with a hard clip operator that will indicate the length of the excised sequence.

secondary_alignment

Whether this alignment is secondary. Equivalent to SAM flag 0x100. A secondary alignment represents an alternative to the primary alignment for this read. Aligners may return secondary alignments if a read can map ambiguously to multiple coordinates in the genome. By convention, each read has one and only one alignment where both secondaryAlignment and supplementaryAlignment are false.