Struct google_genomics1_beta2::ReadGroupExperiment

pub struct ReadGroupExperiment {
    pub sequencing_center: Option<String>,
    pub library_id: Option<String>,
    pub instrument_model: Option<String>,
    pub platform_unit: Option<String>,
}

There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

Fields

sequencing_center

The sequencing center used as part of this experiment.

library_id

The library used as part of this experiment. Note: This is not an actual ID within this repository, but rather an identifier for a library which may be meaningful to some external system.

instrument_model

The instrument model used as part of this experiment. This maps to sequencing technology in BAM.

platform_unit

The platform unit used as part of this experiment e.g. flowcell-barcode.lane for Illumina or slide for SOLiD. Corresponds to the