Struct google_genomics1_beta2::ReadgroupsetMethods

pub struct ReadgroupsetMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on readgroupset resources. It is not used directly, but through the Genomics hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_genomics1_beta2 as genomics1_beta2;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use genomics1_beta2::Genomics;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Genomics::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `align(...)`, `call(...)`, `coveragebuckets_list(...)`, `delete(...)`, `export(...)`, `get(...)`, `import(...)`, `patch(...)`, `search(...)` and `update(...)`
// to build up your call.
let rb = hub.readgroupsets();

Methods

impl<'a, C, A> ReadgroupsetMethods<'a, C, A>

fn export(&self, request: ExportReadGroupSetsRequest) -> ReadgroupsetExportCall<'a, C, A>

Create a builder to help you perform the following task:

Exports read group sets to a BAM file in Google Cloud Storage.

Note that currently there may be some differences between exported BAM files and the original BAM file at the time of import. In particular, comments in the input file header will not be preserved, some custom tags will be converted to strings, and original reference sequence order is not necessarily preserved.

Arguments

request - No description provided.

fn patch(&self, request: ReadGroupSet, read_group_set_id: &str) -> ReadgroupsetPatchCall<'a, C, A>

Create a builder to help you perform the following task:

Updates a read group set. This method supports patch semantics.

Arguments

request - No description provided.

readGroupSetId - The ID of the read group set to be updated. The caller must have WRITE permissions to the dataset associated with this read group set.

fn call(&self, request: CallReadGroupSetsRequest) -> ReadgroupsetCallCall<'a, C, A>

Create a builder to help you perform the following task:

Calls variants on read data from existing read group sets or files from Google Cloud Storage. See the alignment and variant calling documentation for more details.

Arguments

request - No description provided.

fn import(&self, request: ImportReadGroupSetsRequest) -> ReadgroupsetImportCall<'a, C, A>

Create a builder to help you perform the following task:

Creates read group sets by asynchronously importing the provided information.

Note that currently comments in the input file header are not imported and some custom tags will be converted to strings, rather than preserving tag types. The caller must have WRITE permissions to the dataset.

Arguments

request - No description provided.

fn align(&self, request: AlignReadGroupSetsRequest) -> ReadgroupsetAlignCall<'a, C, A>

Create a builder to help you perform the following task:

Aligns read data from existing read group sets or files from Google Cloud Storage. See the alignment and variant calling documentation for more details.

Arguments

request - No description provided.

fn get(&self, read_group_set_id: &str) -> ReadgroupsetGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets a read group set by ID.

Arguments

readGroupSetId - The ID of the read group set.

fn update(&self, request: ReadGroupSet, read_group_set_id: &str) -> ReadgroupsetUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Updates a read group set.

Arguments

request - No description provided.

readGroupSetId - The ID of the read group set to be updated. The caller must have WRITE permissions to the dataset associated with this read group set.

fn coveragebuckets_list(&self, read_group_set_id: &str) -> ReadgroupsetCoveragebucketListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists fixed width coverage buckets for a read group set, each of which correspond to a range of a reference sequence. Each bucket summarizes coverage information across its corresponding genomic range.

Coverage is defined as the number of reads which are aligned to a given base in the reference sequence. Coverage buckets are available at several precomputed bucket widths, enabling retrieval of various coverage 'zoom levels'. The caller must have READ permissions for the target read group set.

Arguments

readGroupSetId - Required. The ID of the read group set over which coverage is requested.

fn delete(&self, read_group_set_id: &str) -> ReadgroupsetDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes a read group set.

Arguments

readGroupSetId - The ID of the read group set to be deleted. The caller must have WRITE permissions to the dataset associated with this read group set.

fn search(&self, request: SearchReadGroupSetsRequest) -> ReadgroupsetSearchCall<'a, C, A>

Create a builder to help you perform the following task:

Searches for read group sets matching the criteria.

Implements GlobalAllianceApi.searchReadGroupSets.

Arguments

request - No description provided.