Struct google_genomics1_beta2::ReferenceSet

pub struct ReferenceSet {
    pub description: Option<String>,
    pub source_uri: Option<String>,
    pub reference_ids: Option<Vec<String>>,
    pub assembly_id: Option<String>,
    pub source_accessions: Option<Vec<String>>,
    pub ncbi_taxon_id: Option<i32>,
    pub id: Option<String>,
    pub md5checksum: Option<String>,
}

A reference set is a set of references which typically comprise a reference assembly for a species, such as GRCh38 which is representative of the human genome. A reference set defines a common coordinate space for comparing reference-aligned experimental data. A reference set contains 1 or more references.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get referencesets (response)

Fields

description

Free text description of this reference set.

source_uri

The URI from which the references were obtained.

reference_ids

The IDs of the reference objects that are part of this set. Reference.md5checksum must be unique within this set.

assembly_id

Public id of this reference set, such as GRCh37.

source_accessions

All known corresponding accession IDs in INSDC (GenBank/ENA/DDBJ) ideally with a version number, for example NC_000001.11.

ncbi_taxon_id

ID from http://www.ncbi.nlm.nih.gov/taxonomy (e.g. 9606->human) indicating the species which this assembly is intended to model. Note that contained references may specify a different ncbiTaxonId, as assemblies may contain reference sequences which do not belong to the modeled species, e.g. EBV in a human reference genome.

id

The Google generated immutable ID of the reference set.

md5checksum

Order-independent MD5 checksum which identifies this reference set. The checksum is computed by sorting all lower case hexidecimal string reference.md5checksum (for all reference in this set) in ascending lexicographic order, concatenating, and taking the MD5 of that value. The resulting value is represented in lower case hexadecimal format.