Struct google_genomics1_beta2::SearchReferenceSetsRequest

pub struct SearchReferenceSetsRequest {
    pub md5checksums: Option<Vec<String>>,
    pub assembly_id: Option<String>,
    pub accessions: Option<Vec<String>>,
    pub page_size: Option<i32>,
    pub page_token: Option<String>,
}

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

search referencesets (request)

Fields

md5checksums

If present, return references for which the md5checksum matches. See ReferenceSet.md5checksum for details.

assembly_id

If present, return reference sets for which a substring of their assemblyId matches this string (case insensitive).

accessions

If present, return references for which the accession matches any of these strings. Best to give a version number, for example GCF_000001405.26. If only the main accession number is given then all records with that main accession will be returned, whichever version. Note that different versions will have different sequences.

page_size

Specifies the maximum number of results to return in a single page.

page_token

The continuation token, which is used to page through large result sets. To get the next page of results, set this parameter to the value of nextPageToken from the previous response.