Struct google_genomics1_beta2::StreamingReadstoreMethods

pub struct StreamingReadstoreMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on streamingReadstore resources. It is not used directly, but through the Genomics hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_genomics1_beta2 as genomics1_beta2;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use genomics1_beta2::Genomics;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Genomics::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `streamreads(...)`
// to build up your call.
let rb = hub.streaming_readstore();

Methods

impl<'a, C, A> StreamingReadstoreMethods<'a, C, A>

fn streamreads(&self, request: StreamReadsRequest) -> StreamingReadstoreStreamreadCall<'a, C, A>

Create a builder to help you perform the following task:

Gets a stream of reads for one or more read group sets. Reads search operates over a genomic coordinate space of reference sequence & position defined over the reference sequences to which the requested read group sets are aligned.

If a target positional range is specified, all reads whose alignment to the reference genome overlap the range are returned.

All reads returned are ordered by genomic coordinate (reference sequence & position). Reads with equivalent genomic coordinates are returned in a deterministic order.

Arguments

request - No description provided.