Struct google_genomics1_beta2::Variant

pub struct Variant {
    pub info: Option<HashMap<String, Vec<String>>>,
    pub variant_set_id: Option<String>,
    pub end: Option<String>,
    pub calls: Option<Vec<Call>>,
    pub created: Option<String>,
    pub id: Option<String>,
    pub filter: Option<Vec<String>>,
    pub start: Option<String>,
    pub names: Option<Vec<String>>,
    pub alternate_bases: Option<Vec<String>>,
    pub reference_name: Option<String>,
    pub quality: Option<f64>,
    pub reference_bases: Option<String>,
}

A variant represents a change in DNA sequence relative to a reference sequence. For example, a variant could represent a SNP or an insertion. Variants belong to a variant set. Each of the calls on a variant represent a determination of genotype with respect to that variant. For example, a call might assign probability of 0.32 to the occurrence of a SNP named rs1234 in a sample named NA12345. A call belongs to a call set, which contains related calls typically from one sample.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get variants (response)

search variants (none)

update variants (request|response)

delete variants (none)

create variants (request|response)

Fields

info

A map of additional variant information.

variant_set_id

The ID of the variant set this variant belongs to.

end

The end position (0-based) of this variant. This corresponds to the first base after the last base in the reference allele. So, the length of the reference allele is (end - start). This is useful for variants that don't explicitly give alternate bases, for example large deletions.

calls

The variant calls for this particular variant. Each one represents the determination of genotype with respect to this variant.

created

The date this variant was created, in milliseconds from the epoch.

id

The Google generated ID of the variant, immutable.

filter

A list of filters (normally quality filters) this variant has failed. PASS indicates this variant has passed all filters.

start

The position at which this variant occurs (0-based). This corresponds to the first base of the string of reference bases.

names

Names for the variant, for example a RefSNP ID.

alternate_bases

The bases that appear instead of the reference bases.

reference_name

The reference on which this variant occurs. (such as chr20 or X)

quality

A measure of how likely this variant is to be real. A higher value is better.

reference_bases

The reference bases for this variant. They start at the given position.