Struct google_genomics1_beta2::VariantsetMethods

pub struct VariantsetMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on variantset resources. It is not used directly, but through the Genomics hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_genomics1_beta2 as genomics1_beta2;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use genomics1_beta2::Genomics;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Genomics::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `delete(...)`, `export(...)`, `get(...)`, `import_variants(...)`, `merge_variants(...)`, `patch(...)`, `search(...)` and `update(...)`
// to build up your call.
let rb = hub.variantsets();

Methods

impl<'a, C, A> VariantsetMethods<'a, C, A>

fn update(&self, request: VariantSet, variant_set_id: &str) -> VariantsetUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Updates a variant set's metadata. All other modifications are silently ignored.

Arguments

request - No description provided.

variantSetId - The ID of the variant to be updated.

fn export(&self, request: ExportVariantSetRequest, variant_set_id: &str) -> VariantsetExportCall<'a, C, A>

Create a builder to help you perform the following task:

Exports variant set data to an external destination.

Arguments

request - No description provided.

variantSetId - Required. The ID of the variant set that contains variant data which should be exported. The caller must have READ access to this variant set.

fn patch(&self, request: VariantSet, variant_set_id: &str) -> VariantsetPatchCall<'a, C, A>

Create a builder to help you perform the following task:

Updates a variant set's metadata. All other modifications are silently ignored. This method supports patch semantics.

Arguments

request - No description provided.

variantSetId - The ID of the variant to be updated.

fn search(&self, request: SearchVariantSetsRequest) -> VariantsetSearchCall<'a, C, A>

Create a builder to help you perform the following task:

Returns a list of all variant sets matching search criteria.

Implements GlobalAllianceApi.searchVariantSets.

Arguments

request - No description provided.

fn get(&self, variant_set_id: &str) -> VariantsetGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets a variant set by ID.

Arguments

variantSetId - Required. The ID of the variant set.

fn merge_variants(&self, request: MergeVariantsRequest, variant_set_id: &str) -> VariantsetMergeVariantCall<'a, C, A>

Create a builder to help you perform the following task:

Merges the given variants with existing variants. Each variant will be merged with an existing variant that matches its reference sequence, start, end, reference bases, and alternative bases. If no such variant exists, a new one will be created.

When variants are merged, the call information from the new variant is added to the existing variant, and other fields (such as key/value pairs) are discarded.

Arguments

request - No description provided.

variantSetId - The destination variant set.

fn delete(&self, variant_set_id: &str) -> VariantsetDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes the contents of a variant set. The variant set object is not deleted.

Arguments

variantSetId - The ID of the variant set to be deleted.

fn import_variants(&self, request: ImportVariantsRequest, variant_set_id: &str) -> VariantsetImportVariantCall<'a, C, A>

Create a builder to help you perform the following task:

Creates variant data by asynchronously importing the provided information.

The variants for import will be merged with any existing data and each other according to the behavior of mergeVariants. In particular, this means for merged VCF variants that have conflicting INFO fields, some data will be arbitrarily discarded. As a special case, for single-sample VCF files, QUAL and FILTER fields will be moved to the call level; these are sometimes interpreted in a call-specific context. Imported VCF headers are appended to the metadata already in a variant set.

Arguments

request - No description provided.

variantSetId - Required. The variant set to which variant data should be imported.