Struct google_genomics1_beta2::VariantMethods

pub struct VariantMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on variant resources. It is not used directly, but through the Genomics hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_genomics1_beta2 as genomics1_beta2;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use genomics1_beta2::Genomics;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Genomics::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `create(...)`, `delete(...)`, `get(...)`, `search(...)` and `update(...)`
// to build up your call.
let rb = hub.variants();

Methods

impl<'a, C, A> VariantMethods<'a, C, A>

fn search(&self, request: SearchVariantsRequest) -> VariantSearchCall<'a, C, A>

Create a builder to help you perform the following task:

Gets a list of variants matching the criteria.

Implements GlobalAllianceApi.searchVariants.

Arguments

request - No description provided.

fn delete(&self, variant_id: &str) -> VariantDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes a variant.

Arguments

variantId - The ID of the variant to be deleted.

fn get(&self, variant_id: &str) -> VariantGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets a variant by ID.

Arguments

variantId - The ID of the variant.

fn create(&self, request: Variant) -> VariantCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a new variant.

Arguments

request - No description provided.

fn update(&self, request: Variant, variant_id: &str) -> VariantUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Updates a variant's names and info fields. All other modifications are silently ignored. Returns the modified variant without its calls.

Arguments

request - No description provided.

variantId - The ID of the variant to be updated.